+++ src/UiIconButton.vue
@@ -34,7 +34,6 @@
         <ui-popover
             constain-focus
             ref="dropdown"
-            remove-on-close
 
             :constrain-to-scroll-parent="constrainDropdownToScrollParent"
             :position="dropdownPosition"
@@ -94,7 +93,7 @@ export default {
         },
         dropdownPosition: {
             type: String,
-            default: 'bottom left'
+            default: 'bottom-start'
         },
         constrainDropdownToScrollParent: {
             type: Boolean,